fix: vpc creation fails because of unexpected variables (#110)
VPC creates are failing my guess is that the required provider blocks in the upstream module which is

```terraform
terraform {
  required_version = ">= 0.13.1"

  required_providers {
    aws = {
      source  = "hashicorp/aws"
      version = ">= 3.63"
    }
  }
}
```
is picking up the latest provider version upstream (`5.x.x`). This is causing issues (my guess is move from 4 to 5 had breaking changes). We should be more pessimistic about version upgrades. I changed it to `4.67.x` (only patches will be allowed). `4.67` is the latest in the `4` series and it is working for us. I can't change it to `3.64.x` because downgrading might not be backwards compatible.

Similar issue might happen with other modules as well, so we need to pin their version as well.